+++ examples/saas-dynamic-dashboards/frontend/app/api/getTests/route.ts
@@ -1,0 +1,5 @@
+import { NextResponse } from "next/server";
+import { testData } from "@/lib/testData";
+export const POST = async () => {
+    return NextResponse.json(testData);
+}