+++ CopilotKit/packages/runtime/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @copilotkit/runtime
 
+## 1.8.14-next.2
+
+### Patch Changes
+
+- 1af7333: - docs: updates ag-ui agent interface
+  - @copilotkit/shared@1.8.14-next.2
+
 ## 1.8.14-next.1
 
 ### Patch Changes