+++ CopilotKit/packages/shared/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @copilotkit/shared
 
+## 1.8.14-next.2
+
 ## 1.8.14-next.1
 
 ### Patch Changes